@@ -21,12 +21,14 @@
     },
     "minimum-stability": "stable",
     "require": {
-        "php": "^7.3",
+        "php": "^7.4",
         "ext-json": "*",
-        "symfony/property-access": "^4.3"
+        "ext-mbstring": "*",
+        "symfony/property-access": "^4.4|^5.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^8.4",
-        "symfony/var-dumper": "^4.3"
+        "symfony/var-dumper": "^4.4|^5.0",
+        "php-coveralls/php-coveralls": "^2.2"
     }
 }
